fixed error-code
