+++ src/components/todoform/todoform.scss
@@ -1,0 +1,46 @@
+
+.todoform {
+    margin-top: 3rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .todoform .buttons {
+    margin-bottom: 2rem;
+  }
+  
+  .todoform .buttons button {
+    padding: 0.5rem 1.2rem;
+    border-radius: 20px;
+    cursor: pointer;
+    border: none;
+  }
+  
+  .todoform .buttons button:focus {
+    outline: none;
+  }
+  
+  
+  .todoform ul {
+    list-style: none;
+    display: flex;
+    align-self: flex-start;
+    flex-wrap: wrap;
+    margin-left: 5%;
+  }
+
+@media Screen and (max-width: 640px) {
+    .todoform {
+      margin-top: 2rem;
+    }
+    .todoform ul {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      margin-left: 0;
+      align-self: center;
+    }
+   
+  }
+
