+++ src/components/todoitem/todoitem.scss
@@ -1,0 +1,65 @@
+.card {
+    display: flex;
+    flex-direction: column;
+    background: rgb(234, 236, 236);
+    margin: 0 1rem 1rem 0;
+    height: 8rem;
+    width: 18rem;
+    border-radius: 0.5rem;
+    padding: 1rem;
+  
+    position: relative;
+  }
+  
+  .card textarea {
+    padding: 0.5rem;
+    border-radius: 8px;
+    border: none;
+    background-color: #e1ebfd;
+    resize: none;
+    color: #271c6c;
+    height: 70%;
+  }
+  .card textarea:disabled {
+    background-color: transparent;
+  }
+  
+  .card .btns {
+    position: absolute;
+    bottom: 1rem;
+    right: 1rem;
+  }
+  
+  .card .btns button {
+    border-radius: 20%;
+    border: none;
+    margin: 0 0.6rem;
+    font-size: 1.4rem;
+    cursor: pointer;
+    color: #271c6c;
+    background-color: transparent;
+  }
+  
+  .card .btns button:focus {
+    outline: none;
+  }
+  
+  .completed {
+    position: absolute;
+    right: 0.3rem;
+    top: 0.3rem;
+    background-color: #867bcd;
+    border: 2px solid #272727;
+    color: #e1ebfd;
+  
+    font-size: 0.8rem;
+    padding: 0.3rem 1rem;
+    border-radius: 20px;
+  }
+  
+  @media Screen and (max-width: 640px) {
+  
+    .card {
+      margin-right: 0;
+    }
+  }
