+++ src/redux/__test__/reducer.test.js
@@ -1,0 +1,46 @@
+import reducer, { addTodos } from '../reducer';
+
+
+test('should return the initial state', () => {
+  expect(reducer(undefined, {})).toEqual([
+    {
+      text: 'Use Redux',
+      completed: false,
+      id: 0
+    }
+  ])
+})
+
+test('should handle a todo being added to an empty list', () => {
+  const previousState = []
+  expect(reducer(previousState, addTodos('Run the tests'))).toEqual([
+    {
+      text: 'Run the tests',
+      completed: false,
+      id: 0
+    }
+  ])
+})
+
+test('should handle a todo being added to an existing list', () => {
+  const previousState = [
+    {
+      text: 'Run the tests',
+      completed: true,
+      id: 0
+    }
+  ]
+  expect(reducer(previousState, addTodos('Use Redux'))).toEqual([
+    {
+      text: 'Run the tests',
+      completed: true,
+      id: 0
+    },
+    {
+      text: 'Use Redux',
+      completed: false,
+      id: 1
+    }
+  ])
+})
+
