--- src/style/main.scss
@@ -1,190 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@500&display=swap");
-
-html {
-  line-height: 1.15;
-}
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  font-family: 'Roboto', sans-serif;
-}
-
-body {
-  background: rgba(242, 249, 250, 0.815);
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-attachment: fixed;
-  color: #222;
-}
-
-.App {
-  margin-top: 3rem;
-  display: flex;
-  flex-direction: column;
-}
-
-.App h1 {
-  display: inline;
-  text-align: center;
-  margin-bottom: 2rem;
-  color: #080808;
-
-  text-shadow: 0 0 5px #bfbcdb, 3px -1px 5px #696779;
-}
-
-.addTodos {
-  display: flex;
-  justify-content: center;
-}
-
-.todo-input {
-  min-width: 10rem;
-  width: 30vw;
-
-  max-height: 2.5rem;
-  background-color: rgb(234, 236, 236);
-  border: none;
-  border-radius: 5px;
-  padding: 0.5rem 1rem;
-  margin-right: 1%;
-
-  align-self: center;
-}
-.todo-input:focus {
-  outline: none;
-  border: 2px solid #080808;
-}
-
-.add-btn {
-  padding-bottom: 12px;
-  background-color: #080808;
-  color: #e1ebfd;
-  border-radius: 50%;
-  border: 2px solid #e1ebfd;
-  font-size: 5.5rem;
-  width: 3.5rem;
-  height: 3.5rem;
-  cursor: pointer;
-  box-shadow: 2px 4px 10px #bfbcdb;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.add-btn:focus {
-  outline: none;
-}
-
-.add-btn:hover {
-  background-color: #e1ebfd;
-  color: #080808;
-}
-
-.todoform {
-  margin-top: 3rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.todoform .buttons {
-  margin-bottom: 2rem;
-}
-
-.todoform .buttons button {
-  padding: 0.5rem 1.2rem;
-  border-radius: 20px;
-  cursor: pointer;
-  border: none;
-}
-
-.todoform .buttons button:focus {
-  outline: none;
-}
-
-
-.todoform ul {
-  list-style: none;
-  display: flex;
-  align-self: flex-start;
-  flex-wrap: wrap;
-  margin-left: 5%;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  background: rgb(234, 236, 236);
-  margin: 0 1rem 1rem 0;
-  height: 8rem;
-  width: 18rem;
-  border-radius: 0.5rem;
-  padding: 1rem;
-
-  position: relative;
-}
-
-.card textarea {
-  padding: 0.5rem;
-  border-radius: 8px;
-  border: none;
-  background-color: #e1ebfd;
-  resize: none;
-  color: #271c6c;
-  height: 70%;
-}
-.card textarea:disabled {
-  background-color: transparent;
-}
-
-.card .btns {
-  position: absolute;
-  bottom: 1rem;
-  right: 1rem;
-}
-
-.card .btns button {
-  border-radius: 20%;
-  border: none;
-  margin: 0 0.6rem;
-  font-size: 1.4rem;
-  cursor: pointer;
-  color: #271c6c;
-  background-color: transparent;
-}
-
-.card .btns button:focus {
-  outline: none;
-}
-
-.completed {
-  position: absolute;
-  right: 0.3rem;
-  top: 0.3rem;
-  background-color: #867bcd;
-  border: 2px solid #272727;
-  color: #e1ebfd;
-
-  font-size: 0.8rem;
-  padding: 0.3rem 1rem;
-  border-radius: 20px;
-}
-
-@media Screen and (max-width: 640px) {
-  .todoform {
-    margin-top: 2rem;
-  }
-  .todoform ul {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-left: 0;
-    align-self: center;
-  }
-  .card {
-    margin-right: 0;
-  }
-}
